feature: move fvm to the lhs and fvc to rhs...
(cdiFoam): in concentration and donnan potential equation, move the fvm
expressions in the source term to the right hand side of the equation.
for that i create a fvscalarfield out of the coefficient and shove that
into the `fvm::ddt`'s first argument, so that rhs becomes something
like:
    fvm::ddt(1+c_coeff, c)
i did this for the Donnan potential equation too. The result was, the
average outlet concentration came almost down to 17.